fix: skip backend resolution for the remaining host-only verbs
network print and diff only render policy, ssh-init copies files,
validate-extensions reads specs, and devcontainer generate resolves the build
config without an engine, yet all of them detected the backend and could ask
the engine question on a dual-engine host. The predicate moves next to
isRunAction, where the other action predicates live.
